Update desktoprenamer extension
- chore: add eslint.config.js for ESLint v9 flat config compatibility
- style: improve code styles
- refactor: extract duplicate space fetching and RenameSpaceForm to shared module
- fix(batch-move): avoid double error toast on failure
