+++ extensions/desktoprenamer/eslint.config.js
@@ -1,6 +1,5 @@
-const { defineConfig } = require("eslint/config");
 const raycastConfig = require("@raycast/eslint-config");
 
-module.exports = defineConfig([
-  ...raycastConfig,
-]);
+module.exports = [
+  ...raycastConfig.flat(Infinity),
+];
